requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==2.0.12
dewiki==0.1.0
distlib==0.3.8
Django==5.0.7
filelock==3.15.4
idna==3.7
path @ git+https://github.com/jaraco/path.git@34634dc494d8385c0b40a3bedcce5458983b3a5d
path.py==12.5.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.27.1
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==1.26.19
virtualenv==20.26.3
Wikipedia-API==0.6.0
